pyproject.toml
[project]
name = "llm-notify-mcp"
version = "0.1.0"
description = "Local notification bridge for LLM agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pync>=2.0.3",
    "pyyaml>=6.0.1",
    "aiofiles>=23.2.1",
    "aiohttp>=3.8.0",
    "mcp>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.2",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "pre-commit>=3.4.0",
    "httpx>=0.24.0",
]

[project.scripts]
llm-notify-mcp = "llm_notify_mcp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_notify_mcp"]
